Refactor clientevent structure
* Move structure outside of client header
